importa_csv/requirements.txt
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
